update SG/MG C++ testing infrastructures

+++ cpp/src/experimental/renumber_edgelist.cu
@@ -544,7 +544,8 @@ renumber_edgelist(raft::handle_t const& handle,
 
   for (size_t i = 0; i < edgelist_major_vertices.size(); ++i) {
     rmm::device_uvector<vertex_t> renumber_map_major_labels(
-      col_comm_rank == i ? vertex_t{0} : partition.get_matrix_partition_major_size(i),
+      col_comm_rank == static_cast<int>(i) ? vertex_t{0}
+                                           : partition.get_matrix_partition_major_size(i),
       handle.get_stream());
     device_bcast(col_comm,
                  renumber_map_labels.data(),
@@ -563,7 +564,8 @@ renumber_edgelist(raft::handle_t const& handle,
       invalid_vertex_id<vertex_t>::value};
     auto pair_first = thrust::make_transform_iterator(
       thrust::make_zip_iterator(thrust::make_tuple(
-        col_comm_rank == i ? renumber_map_labels.begin() : renumber_map_major_labels.begin(),
+        col_comm_rank == static_cast<int>(i) ? renumber_map_labels.begin()
+                                             : renumber_map_major_labels.begin(),
         thrust::make_counting_iterator(partition.get_matrix_partition_major_first(i)))),
       [] __device__(auto val) {
         return thrust::make_pair(thrust::get<0>(val), thrust::get<1>(val));
